+++ source/GLG3Dcpp/PosedModel.cpp
@@ -44,7 +44,7 @@ void PosedModel::sort(
     Array<ModelSorter> op;
     Array<ModelSorter> tr;
     
-    for (int m = 0; m < inModels.size(); ++m) {
+    for (size_t m = 0; m < inModels.size(); ++m) {
         if (inModels[m]->hasTransparency()) {
             tr.append(ModelSorter(inModels[m], wsLook));
         } else {
@@ -57,12 +57,12 @@ void PosedModel::sort(
     op.sort(SORT_INCREASING);
 
     transparent.resize(tr.size(), DONT_SHRINK_UNDERLYING_ARRAY);
-    for (int m = 0; m < tr.size(); ++m) {
+    for (size_t m = 0; m < tr.size(); ++m) {
         transparent[m] = tr[m].model;
     }
 
     opaque.resize(op.size(), DONT_SHRINK_UNDERLYING_ARRAY);
-    for (int m = 0; m < op.size(); ++m) {
+    for (size_t m = 0; m < op.size(); ++m) {
         opaque[m] = op[m].model;
     }
 }
